requirements.txt
# Tested on Python 3.9.0

# Unittest
coverage
pytest

# Code analysis
bandit
#metrics.pylint
flake8
flake8-docstrings
flake8-polyfill
pylint
radon

# Documentation
pdoc3
sphinx

# Pythonic Interface
python-interface
